requirements.txt
ixbrlparse
psycopg2-binary
requests
python-dotenv
psutil
openpyxl
pandas
PyYAML
